Issue-163: use module instead of inheritance for common context logic
