+++ lib/holidays/use_case/context/between.rb
@@ -1,7 +1,9 @@
 module Holidays
   module UseCase
     module Context
-      class Between < ContextCommon
+      class Between
+        include ContextCommon
+
         def initialize(holidays_by_month_repo, day_of_month_calculator, custom_methods_repo, proc_result_cache_repo)
           @holidays_by_month_repo = holidays_by_month_repo
           @day_of_month_calculator = day_of_month_calculator
@@ -13,15 +15,17 @@ module Holidays
           validate!(start_date, end_date, dates_driver, regions)
 
           holidays = []
-          
           holidays = make_date_array(dates_driver, regions, observed, informal)
-
           holidays = holidays.select{|holiday|holiday[:date].between?(start_date, end_date)}
-          
           holidays.sort{|a, b| a[:date] <=> b[:date] }
         end
 
         private
+
+        attr_reader :holidays_by_month_repo,
+                    :day_of_month_calculator,
+                    :custom_methods_repo,
+                    :proc_result_cache_repo
 
         def validate!(start_date, end_date, dates_driver, regions)
           raise ArgumentError unless start_date
@@ -35,7 +39,6 @@ module Holidays
 
           raise ArgumentError if regions.nil? || regions.empty?
         end
-
       end
     end
   end
